requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
botocore==1.42.30
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
idna==3.11
isort==7.0.0
jmespath==1.0.1
pycparser==2.23
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
apache-airflow-providers-amazon